# BE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SWAGGER for API's of  [gauriz/ng9-node-user-access](https://github.com/gauriz/ng9-node-user-access) ",
        "version": "1.0.0",
        "title": "ng9-node-user-access",
        "contact": {
            "email": "gauri.x3@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3000",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/user-logs": {
            "get": {
                "tags": [
                    "user logs"
                ],
                "summary": "get all user logs from the system",
                "description": "",
                "operationId": "userlogs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sortKey",
                        "in": "query",
                        "description": "Sort Key",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "sortDir",
                        "in": "query",
                        "description": "Sort Direcion",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "userName",
                        "in": "query",
                        "description": "User Name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/user-log-model"
                        }
                    },
                    "401": {
                        "description": "No authorization provided or Invalid Token!",
                        "schema": {
                            "$ref": "#/definitions/error-model"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "error-model": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "user-log-model": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "login_time": {
                    "type": "string"
                }
            }
        }
    },
...
}